feat: active/inactive styling for dimension items
The callback is passed from the app that uses the DataDimension
component and is called for each data dimension item for deciding if
render it as active or inactive.
The item will still be draggable like the others and the selectedItems
list is handled as before (all the items are used).
DV app and d2ca have the logic for using only the first item in the list
for single value visualization type.
The feature here is basically just applying a different style to the
item component to visually differentiate the one actually used from the
rest.
By reordering the selected items (right hand side of the selector) is
possible to change the item used by the single value visualization.