fix: latest Highcharts "No data to display" works again
The no-data-to-display module in Highcharts is now imported directly
from the Highcharts package, instead of using a separate npm module
(same with the other Highcharts modules used).
For missing data in the analytics response we add a null value to the
serie's array to maintain the correct positions for the values.
In case of a whole serie without data however, we passed an array filled
with null values to Highcharts.
The latest Highcharts no-data-to-display version has an additional check
which caused the "No data to display" message to not show in this
circumstance.
The fix solves this problem; if a whole serie is empty, an empty array
is returned instead.