fix: add function to replace a dimension in the layout (#79)
The specific use case that caused the creation of this function:

In the past, when a visualization layout is saved and it contains org unit levels, the level number (e.g., 1, 2, 3, 4) was saved. In the new DV, we now store the uid of the org unit level, but for historical data, the layout will still contain the numeric values. To soft-migrate old data to the uid, we need to replace the ou dimension in the layout.

An additional function was added: axisHasOuDimension, to parallel the other related functions (e.g., axisHasPeriodDimension).

[DHIS2-6990]