fix: axisItemHandler to pass new props (#294)
* fix: include new prop requireItems
* fix: updated the fn signature and introduced the breaking change

Because of eslint max-params, the axisItemHandler signature in DV had to be changed from 3 regular params to an object with 5 props, which causes a breaking change. As per commit 2a69679 axisItemHandler requires an object param with 5 props instead of just 3 regular params. To be compatible with this version, change axisItemHandler params to an object and pass in requireItems and isDimensionInLayout.

BREAKING CHANGE:
axisItemHandler requires an object with 5 props instead of 3 regular props