chore: move PeriodSelector component to analytics [DHIS2-8638] (#394)
Part of fix for DHIS2-8638

The PeriodSelector component has been part of a circular dependency between d2-ui and analytics. This has created a lot of headaches with upgrading and other at-times strange errors and behaviour. In the future, PeriodSelector will be part of ui-widgets, but until that can happen, Period Selector is being moved to analytics to resolve the circular dependency issues

* No functional changes
* Code is nearly identical to what was in d2-ui. Only changes were related to converting the css file to styled-jsx.