feat: introduce fixed-period-select component (#509)
This introduces a new component `FixedPeriodSelect` component which is a variation on the `PeriodTransfer` for when a user only needs to select a single period. It is a controlled input, so can be used with any from-management library. Demo stories and unit tests are included.

The internal component previously called `PeriodSelector` has been renamed to `PeriodTransfer`, because this better reflects the fact that it renders a `ui-core` `Transfer` component, and also because this distinguishes it more clearly from the new FixedPeriodSelect. Since this component was never exported from the published lib, this doesn't cause any breaking changes.