fix: extract color from pattern for trendline
When the colorSet option is set to patterns, there isn't a color passed
to highcharts, but the index of the pattern.
The trendline tho needs a color to distinguish between the different
trendlines when using multiple data items.
Luckily the Highcharts patterns have a color in their definition, the
same one used in the pattern.