fix: fix hide empty categories with non-available data DHIS2-8174
The bug caused the hide empty categories option in DV to not work when
one or more of the dx dimensions had no data (ie. empty array in the
response from analytics).

Empty data caused undefined values to be added to the list checked for
all null values; with this check failing, the index in the category list
was never added to the list of empty category indexes, thus the chart
was not "filtered".

(cherry picked from commit 5703a570ecbf31c27cbe0a2b327723fcb4d3e231)