fix: provide default values for visualizations (#947)
Fixes DHIS2-11304

Provides default values for fontSize and displayDensity, which are missing from the visualization when it is originally a Map. For instance, in dashboard, a Map can be viewed as a Table. Since those properties were missing, the table had an unexpected font size and display density (see screenshots)