fix: provide default props for maps being displayed as table [16.x] (#980)
Fixes DHIS2-11304

Backport of #947

Key features
Provides default values for fontSize and displayDensity, which are missing from the visualization when it is originally a Map. For instance, in dashboard, a Map can be viewed as a Table. Since those properties were missing, the table had an unexpected font size and display density (see screenshots from #947)