fix: remove unnecessary or condition
This was causing top and left CSS properties to be added to the column
header cells even when fixedRowHeaders only was enabled.
Because the position: sticky rule was not set in that case, the actual
behaviour of the table when scrolling is the same as when not setting
top/left on the header column cells.