refactor: fetch necessary data and update subscribe button directly
This works better than having to rely on the app to update a prop in
order to render the unit's content.
For example, the subscribe/unsubscribe state depends on a field in the
AO, that would mean the app has to refetch the AO just for that field
which is not even used by the app. That refetch can be costly in terms
of re-renders of various components in the app.