feat: disable Save button when onSave not passed (TECH-1013)
We need a way of disabling the Save button in certain situations.
The logic should be in the consumer app.
The idea is to leverage the onSave prop and use it to enable/disable the
Save button.
If the callback is not passed disable the Save button.

(cherry picked from commit fa87c8b9ef3fe8e237e7d21423c81daf0ff6c817)