fix: avoid canceled request error
There is an edge case where the Modal is rendered after the translations
are saved before it's closed.
This re-render causes a useEffect to run where a request to i18n is
performed.
This request is canceled as soon as the Modal is closed, throwing an
error that is shown in the AlertBar.

This request can be avoided if the Modal is closed before the callback
after saving the translations is called.

This issue happened in LL app.