fix: do not pass translation object when value is empty
This causes a 409 response from the translations API.
If the input field is cleared, assuming the user wants to remove a
translation for a language/string entirely, do not pass the object with
an empty value, but remove it entirely from the list of translations.
We still need to pass the full list of translations because the API does
not support PATCH requests so the logic is a bit more complex.