fix: use page number from response
This is to avoid the edge case where a page does not exist anymore after
deleting the visualization that was listed in it.

The page in the component state needs to be set to the page returned in
the response when it differs.

Unfortunately there is an extra request with the not existing page,
because until we get the response from the api we don't know if a page
has been removed or not as a consequence of deleting visualizations.