feat: dnd for custom calculations (#1416)
dnd for calculations

Some technical notes:

* DndContext is mostly copied from Line-list.
The only thing that has really changed is the handleDragEnd function.

* When an item is added to the formula, it is assigned an id based on the id
or label of the item and a counter. Dnd requires that each draggable item have an id.

* In the INPUT formula item, there is a setTimeout that was needed in order to get
the cursor to stay in the input element once it was added to the formula.

* In the FormulaField, there is a special, invisible drop zone always present in 
the formula field. It makes it possible to drop an item before the first item. 
This is the DropZone component. This is almost identical to what we have in LL

* Also in FormulaField, there is a little trickery that makes it possible to show
the full drop indicator on top of the border.
This is using the <div className="border> element.
The problem was that with overflow set on the formula field,
the drop indicator is getting cut off. An alternative to the current solution would be
to increase the padding inside FormulaField to leave enough space
for the full drop indicator to show, but that means a padding of 12px
which looked strange to me. But it's worth discussing.