fix: default visualizationHasTimeDimension to true
Without this default value the changes in this PR would actually
constitute to a breaking change, at least for the line-listing-app:
if the `visualizationHasTimeDimension` prop is not provided and the
`type` equals `eventVisualization`, the helptext would be rendered.
By defaulting to `true` we avoid this, only explicitly setting
`visualizationHasTimeDimension` to `false` will cause the help text to
show.

I think this is mainly a theoretical issue, because the line-listing-app
will always set the prop. But who knows: there could be a 3rd party app
out there using the analytics library which would break because of this.