chore: new github workflow that copies build of all non-prod branches to d2-ci (#1676)
Every branch that is not listed as one of the prod branches (from node-publish.yml)
will now have it's build pushed to d2-ci for every pushed commit.
We can then use these builds during development and testing.

Changes to README

*remove build badges for versions that are no longer supported
*add docs about how to use the build
New gh workflow file

after building the lib, use yarn pack to create the correct structure of the packaged lib.
Then unzip the result and push that to d2-ci, which can then be installed by the consuming apps.