fix: return new name and desc from file menu rename instead of PATCH request [DHIS2-19433] (#1771)
Fixes DHIS2-19433

BREAKING CHANGE:
Instead of using patch, which causes several backend errors, the rename dialog
now just returns the user's changes to name and description, and leaves it
to the app to save the changes.

To support save centrally for analytics apps, preparePayloadForSave was added

This is a workaround until problems with patch are fixed.