fix: fix sorting in PT when percentage of rows is used (DHIS2-5455) (#1774)
The sorting always used the raw value.
This breaks the sorting when percentage of row is used because the
rendered value is computed comparing cell values across rows, but the
sorting is always by column.
For this scenario, the computed percentage value should be used in the
sorting function instead.