fix: do not fetch subscribers as that should be fetched already (#1777)
Implements DHIS2-19506

BREAKING CHANGE:

As a quick workaround for renaming bugs, a solution was implemented where
the preparePayloadForSave function would fetch the subscribers property
so that it would not be overwritten during a rename. However, it is still possible 
to overwrite other newer changes by a different user. The best solution to deal
with the lack of a PATCH is that each app fetches the entire visualization
object prior to rename. Thus, the only job for the preparePayloadForSave
function is to handle name and description changes.