fix: remove overflow-y which breaks the EVER plugin content
This CSS rule was initially added for fixing DHIS2-15558 and it seems to
only affect the LL plugin where a vertical scrollbar appears in the
Modal.
Although the LL app is going to be replaced by the EVER app,
there might be a better fix for this which works in both apps.