feat: add InterpretationsProvider and refactor components to use it (#1791)
* feat: add InterpretationsProvider and refactor all components to use it

* fix: export InterpretationsProvider

* test: add unit tests for new interpretation-manager hooks

* chore: move mock data so it's not seen as a test

* refactor: let useInterpretationsList return IDs grouped by date

* chore: fix interpretation unit story

* chore: address sonarqube suggestions

* fix: pass id as object to hook

* feat: add components for dashboard-app

* chore: make InterpretationThread internal

BREAKING CHANGE: apps can no longer import InterpretationThread

* fix: remove overflow-y which breaks the EVER plugin content

This CSS rule was initially added for fixing DHIS2-15558 and it seems to
only affect the LL plugin where a vertical scrollbar appears in the
Modal.
Although the LL app is going to be replaced by the EVER app,
there might be a better fix for this which works in both apps.

* Revert "fix: remove overflow-y which breaks the EVER plugin content"

We should fix the issue in the EVER LineList components instead.
The change also causes an extra vertical scrollbar when there is a lot
of content in the interpretations thread.

* chore: address PR comments

* chore: remove redundant prop from interpretations unit in demo story

* fix: memoize the InterpretationThread to fix initial focus issues

* Revert "fix: memoize the InterpretationThread to fix initial focus issues"

This reverts commit b96a3e2faf27bdc4d14c33879f397d59cc38527f.

* fix: focus the rich-text-editor after a timeout

---------

Co-authored-by: Edoardo Sabadelli <edoardo@dhis2.org>